Restructure srs directory.
