+++ .gitignore
@@ -3,4 +3,4 @@ npm-debug.log
 *.pem
 .env
 coverage
-lib
+/lib
