+++ src/lib/commands.ts
@@ -1,0 +1,8 @@
+// import { COMMAND_STATUS } from 'types'
+//
+// export const handleCommands(comment : string) : COMMAND_STATUS => {
+//   let result : COMMAND_STATUS = COMMAND_STATUS.FAIL;
+//
+//
+//   return result
+// }
